fix: 환경 변수 설정

+++ socket.js
@@ -23,12 +23,14 @@ const e = require('express');
 const { table } = require('console');
 
 // db 연결 확인
-DB.sequelize
-  .sync({ force: false })
-  .then(() => {
-    console.log('database 연결 성공');
-  })
-  .catch(console.error);
+if (process.env.NODE_ENV === 'development') {
+  DB.sequelize
+    .sync({ force: false })
+    .then(() => {
+      console.log('database 연결 성공');
+    })
+    .catch(console.error);
+}
 
 const io = new Server(server, {
   cors: {
